change classes names, improve pixel list, remove color
